requirements.txt
requests==2.32.3
beautifulsoup4==4.13.4 
tokenizers==0.21.1
psycopg2-binary==2.9.10
pika==1.3.2
